refactor : swagger 관련 설정 파일 위치 이동 (#122)
