+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/bakery/controller/BakeryController.java
@@ -5,7 +5,8 @@ import com.breaditnow.customer.bakery.controller.res.BakeryDetailResponse;
 import com.breaditnow.customer.bakery.controller.res.HotBakeryPageResponse;
 import com.breaditnow.customer.bakery.service.BakeryPageService;
 import com.breaditnow.customer.bakery.service.BakeryService;
-import com.breaditnow.customer.common.req.GeoPointRequest;
+import com.breaditnow.customer.common.application.request.GeoPointRequest;
+import com.breaditnow.customer.common.presentation.swagger.docs.BakeryControllerDocs;
 import com.breaditnow.customer.common.security.annotation.AuthCustomer;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
