+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/bakery/controller/BakeryFavoriteController.java
@@ -4,7 +4,8 @@ import com.breaditnow.common.response.ApiSuccessResponse;
 import com.breaditnow.customer.bakery.controller.res.BakeryFavoritePageResponse;
 import com.breaditnow.customer.bakery.service.BakeryFavoritePageService;
 import com.breaditnow.customer.bakery.service.BakeryFavoriteService;
-import com.breaditnow.customer.common.req.GeoPointRequest;
+import com.breaditnow.customer.common.application.request.GeoPointRequest;
+import com.breaditnow.customer.common.presentation.swagger.docs.BakeryFavoriteControllerDocs;
 import com.breaditnow.customer.common.security.annotation.AuthCustomer;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
