+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/common/presentation/swagger/annotation/CommonErrorCodeExamples.java
@@ -1,4 +1,4 @@
-package com.breaditnow.customer.common.swagger.annotation;
+package com.breaditnow.customer.common.presentation.swagger.annotation;
 
 import static java.lang.annotation.ElementType.*;
 import static java.lang.annotation.RetentionPolicy.*;
@@ -6,10 +6,10 @@ import static java.lang.annotation.RetentionPolicy.*;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import com.breaditnow.customer.common.exception.CustomerErrorCode;
+import com.breaditnow.common.exception.CommonErrorCode;
 
 @Target(METHOD)
 @Retention(RUNTIME)
-public @interface CustomerApiErrorCodeExamples {
-	CustomerErrorCode[] value();
+public @interface CommonErrorCodeExamples {
+	CommonErrorCode[] value();
 }
