+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/common/presentation/swagger/annotation/ExternalErrorCodeExamples.java
@@ -1,4 +1,4 @@
-package com.breaditnow.customer.common.swagger.annotation;
+package com.breaditnow.customer.common.presentation.swagger.annotation;
 
 import static java.lang.annotation.ElementType.*;
 import static java.lang.annotation.RetentionPolicy.*;
