+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/common/presentation/swagger/docs/BakeryControllerDocs.java
@@ -1,10 +1,10 @@
-package com.breaditnow.customer.bakery.controller;
+package com.breaditnow.customer.common.presentation.swagger.docs;
 
 import com.breaditnow.common.response.ApiSuccessResponse;
 import com.breaditnow.customer.bakery.controller.res.BakeryDetailResponse;
 import com.breaditnow.customer.bakery.controller.res.HotBakeryPageResponse;
-import com.breaditnow.customer.common.req.GeoPointRequest;
-import com.breaditnow.customer.common.swagger.annotation.DomainErrorCodeExamples;
+import com.breaditnow.customer.common.application.request.GeoPointRequest;
+import com.breaditnow.customer.common.presentation.swagger.annotation.DomainErrorCodeExamples;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.Parameters;
