+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/customer/domain/Customer.java
@@ -7,8 +7,8 @@ import lombok.Getter;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.multipart.MultipartFile;
 
-import static com.breaditnow.customer.common.core.ValidationUtils.requireValid;
-import static com.breaditnow.customer.common.core.ValidationUtils.setIfNotNull;
+import static com.breaditnow.customer.common.domain.ValidationUtils.requireValid;
+import static com.breaditnow.customer.common.domain.ValidationUtils.setIfNotNull;
 import static com.breaditnow.customer.common.exception.CustomerErrorCode.*;
 
 @Getter
