+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/notification/controller/NotificationController.java
@@ -5,6 +5,7 @@ import static org.springframework.data.domain.Sort.Direction.*;
 import java.util.List;
 import java.util.Map;
 
+import com.breaditnow.customer.common.presentation.swagger.docs.NotificationControllerDocs;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.DeleteMapping;
